UAS KELOMPOK 3/face-recognition-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.1.0
scikit-learn==1.3.2  # Untuk klasifikasi wajah